refactor: extract @kolu/terminal-awareness from kolu-server (P1a)
Lift the per-terminal provider set (git · PR · agent · foreground) and the
generic awareness schemas out of kolu-server/kolu-common into a new standalone
leaf package, @kolu/terminal-awareness. Behaviour-preserving: kolu-server still
runs the providers in-process and writes terminalMetadata exactly as before —
only where the code lives changes (one shared copy, never a duplicate).

- providers.ts + agentRecency.ts (and its test) move verbatim; the lone host
  coupling (the pino logger) is now injected as a startProviders parameter
  instead of imported from kolu-server.
- The awareness value schema inverts home: @kolu/terminal-awareness/schema OWNS
  AwarenessValue (cwd · git · lastAgentCommand · lastActivityAt · pr · agent ·
  foreground) plus AgentInfo / PrResult / PrUnavailableSource / Foreground /
  TerminalId. kolu-common now EXTENDS that base — ServerPersistedTerminalFields
  = AwarenessPersistedFields + location — and re-exports the moved symbols, so
  the ~17 consumer import sites stay unchanged.
- terminal-awareness depends only on generic packages (anyagent · anyforge ·
  kolu-git/github · the per-agent packages · kaval · @kolu/surface) — no
  kolu-app package — so the future arivu daemon can reuse it over ssh.

First slice (P1a) of the arivu plan (docs/atlas/src/content/atlas/arivu.mdx).
