+++ packages/common/package.json
@@ -30,6 +30,7 @@
     "@kolu/surface": "workspace:*",
     "@kolu/surface-app": "workspace:*",
     "@kolu/surface-daemon-supervisor": "workspace:*",
+    "@kolu/terminal-awareness": "workspace:*",
     "@orpc/contract": "^1.13.13",
     "anyagent": "workspace:*",
     "anyforge": "workspace:*",
