+++ packages/server/package.json
@@ -19,6 +19,7 @@
     "@kolu/surface": "workspace:*",
     "@kolu/surface-app": "workspace:*",
     "@kolu/surface-daemon-supervisor": "workspace:*",
+    "@kolu/terminal-awareness": "workspace:*",
     "@kolu/terminal-protocol": "workspace:*",
     "@orpc/client": "^1.13.13",
     "@orpc/contract": "^1.13.13",
