+++ packages/terminal-awareness/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@kolu/terminal-awareness",
+  "version": "0.1.0",
+  "private": true,
+  "license": "AGPL-3.0-or-later",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./schema": "./src/schema.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "test:unit": "vitest run"
+  },
+  "dependencies": {
+    "@kolu/surface": "workspace:*",
+    "anyagent": "workspace:*",
+    "anyforge": "workspace:*",
+    "kaval": "workspace:*",
+    "kolu-claude-code": "workspace:*",
+    "kolu-codex": "workspace:*",
+    "kolu-git": "workspace:*",
+    "kolu-github": "workspace:*",
+    "kolu-opencode": "workspace:*",
+    "pino": "^10.3.1",
+    "ts-pattern": "^5.9.0",
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "typescript": "^5.8.0",
+    "vitest": "^4.1.0"
+  }
+}
