+++ packages/terminal-awareness/src/index.ts
@@ -1,0 +1,14 @@
+/**
+ * `@kolu/terminal-awareness` — the per-terminal awareness provider set (git ·
+ * PR · agent · foreground) plus the generic `AwarenessValue` schema it
+ * produces, lifted out of kolu-server so a standalone daemon (`arivu`) and
+ * kolu-server share ONE copy of the freshness-critical provider code.
+ *
+ * The package names no kolu-app package: its lone host coupling — a logger —
+ * is injected as a `startProviders` parameter. Consumers that only need the
+ * schemas (no providers, no node/kaval runtime) import `./schema` directly.
+ */
+
+export * from "./agentRecency.ts";
+export * from "./providers.ts";
+export * from "./schema.ts";
